+++ framework/__init__.py
@@ -1,4 +1,5 @@
 from .support.buttons import Button, Btn, Buttons
+from .support.filters import private, group, supergroup, text, photo, video, document
 from .support.screen import screen, ScreenResponse, data, _1, _2, _3, _4, _5
 from .support.Log import Log
 
@@ -15,4 +16,11 @@ __all__ = [
     "_3",
     "_4",
     "_5",
+    "private",
+    "group",
+    "supergroup",
+    "text",
+    "photo",
+    "video",
+    "document",
 ]