+++ framework/support/__init__.py
@@ -1,9 +1,18 @@
 from .buttons import Button, Btn, Buttons
 from .Log import Log
+from .filters import FilterProxy, private, group, supergroup, text, photo, video, document
 
 __all__ = [
     "Button",
     "Btn",
     "Buttons",
     "Log",
+    "FilterProxy",
+    "private",
+    "group",
+    "supergroup",
+    "text",
+    "photo",
+    "video",
+    "document",
 ]